// resources/pathname.ts
export enum Pathname {
  Home = '/',
  Example = '/example',
}

export enum BlogSlug {
  WavePlane = 'wave-plane',
  NextJsShaderSetup = 'nextjs-setup-glsl-shaders',
  ImageSequenceHeader = 'scroll-driven-image-sequence-header',
  // AppDevelopmentGuide = 'app-development-guide',
  AnimatedCSSGrid = 'animated-css-grid-with-tailwind-gsap',
  // EnergyTransfer = 'energy-transfer-threejs',
  // ScatteredMeshPoints = 'scattered-mesh-points-fbo-simulation-react-three-fiber',
}

export enum ExampleSlug {
  ImageSequence = 'scroll-driven-image-sequence-header',
  ScrollingBackgroundShader = 'scrolling-background-shader',
  ScrollingThreeJs = 'three-scrolling-scene',
  StarsParticles = 'three-stars-particles',
  FBOEffects = 'fbo-effects',
  InfiniteMarquee = 'infinite-scrolling-marquee',
  WavePlane = 'three-wave-plane',
  RayMarching = 'raymarching-fragment-shader',
  AnimatedCSSGrid = 'animated-grid-tailwind-gsap',
  EnergyTransfer = 'three-energy-transfer-concept',
  LoopPoints = 'three-loop-mesh-fbo-points',
}